[#29] - Store and load email
