+++ package.json
@@ -14,6 +14,7 @@
     "@fortawesome/fontawesome-svg-core": "1.2.36",
     "@fortawesome/free-solid-svg-icons": "5.15.4",
     "@fortawesome/react-native-fontawesome": "0.2.7",
+    "@react-native-async-storage/async-storage": "^1.15.14",
     "@react-native-firebase/app": "12.9.3",
     "@react-native-firebase/auth": "12.9.3",
     "@react-native-firebase/firestore": "12.9.3",
@@ -31,7 +32,8 @@
     "react-native-screens": "3.8.0",
     "react-native-svg": "12.1.1",
     "react-redux": "7.2.5",
-    "redux-saga": "1.1.3"
+    "redux-saga": "1.1.3",
+    "typed-async-storage": "^3.1.2"
   },
   "devDependencies": {
     "@babel/core": "7.12.9",
